backend/requirements.txt
fastapi
uvicorn
openai
numpy
scikit-learn
python-dotenv
spacy
httpx

# Install spaCy English model
# After installing requirements, run:
# python -m spacy download en_core_web_sm 